fullstack-example/backend/requirements.txt
fastapi
uvicorn[standard]
pyyaml
azure-search-documents
openai
python-dotenv
pydantic
